rename users schema and dal
